services/platform-api/pyproject.toml
[project]
name = "platform-api"
version = "0.1.0"
description = "Metadata API"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "duckdb",
    "pyyaml",
    "pydantic",
    "httpx",
    "langchain-ollama",
    "langchain-core",
    "boring-semantic-layer[agent]",
]
